requirements.txt
Flask==2.3.2
opencv-python==4.8.0.74
Pillow==10.0.0
numpy==1.24.3
torch==2.0.1
transformers==4.33.1
requests==2.31.0